HYD-6796 Update to Lustre build version 2.7.19.2
The Lustre team has provided us with the RC2 build for EE 3.1, which
needs integrated into overall EE. This is the first Lustre build based
on RHEL 7.3. We also need to use the latest e2fsprogs with it.

Update to the Lustre version 2.7.19.2 and the latest e2fsprogs by
changing the following build parameters:

  JENKINS_BUILD_NUMBER = 150
  E2FSPROGS_VERSION := 1.42.13.wc5

The following chroma-dependencies are affected:
  lustre[-client]
  zfs
  e2fsprogs

  - Update chroma-dependencies/include/Makefile.lustre_build_meta
  - Update chroma-dependencies/include/Makefile.e2fsprogs_version
  - Update references to hardcoded kernel-devel version.

Change-Id: I1cba49c8e3da9d92f83d5c63c7f4396cce70a0f9
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/23903
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>