HYD-6859 Monitored filesystems cannot be forgotten
It is not possible to forget monitored filesystems because
a target is dependent on the PacemakerConfiguration being in
the started state. However PacemakerConfiguration does not exist
for monitored fileSystems.

Simply update the code to ensure only required PacemakerConfiguration
state when a PacemakerConfiguration is present.

Add tests to Forget the Filesystem and MGT at the end of the Monitored
Filesystem Tests.

 - Update the code to ensure only required PacemakerConfiguration
   state when a PacemakerConfiguration is present.
 - Add _forget_filesystem to TestFilesystemDetection and call at the
   end of every test.

Change-Id: I863c31a0839c5c98ed828d6e910e5032ff8adcaf
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23917
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>