HYD-6781 Documentation update about removing datasets from zpools
IML documentation should be updated to explain the following:

Zpools do not appear on the volumes page after the user removes a
filesystem. This is because the zpools still have IML-created datasets
when the filesystem is removed. IML cannot destroy the datasets when the
user removes a filesystem because otherwise the filesystem would not be
recoverable. Users will have to remove the datasets manually (as per ZFS
documentation) before reusing the zpools as volumes in IML.

  - Bump help-docs to 1.2.0.

Change-Id: I2cc9492c22340f4bb98b4601493ae918e145e6da
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/23922
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>