HYD-3406 Enable non-HA mode in IML
Whilst the best known method for Lustre is to use HA mode it is
often the case that customer want to configure some or all of there
system in a non-HA configuration.

Durring the development of 3.0 and to a less extent 3.1 the Manager
was enhanced make the component nature of its design user selectable
meaning that users (via the use of the profiles) can decide which
components of IML they are doing to use.

A profile not installing Syslogging for example or not install the
NTP configuration can easily be created. By extending the GUI this
feature could be extended to allow the user to define the features
they want at install time.

This patch makes use of the facility to provide for a NON-HA server
install. This facility is completely flexible and the user can
create a filesystem of HA and non-HA nodes. Mixing them in any
fashion that they so wish. When create a non-HA node the pacemaker
and corosync HA packages are not installed at all leading to a
quicker install process.

For those tests not requiring HA we now make use of the non-HA
install which speeds up the testing significantly.

Finally the mirrored version of fence-agents changed from 4.0.11-27
to 4.0.11-47 (it cost most of my Christmas / New Year break to work
out that small detail and so I've updated the value in this patch -
not complaining because I reviewed that last change that allow the
version to be pinned). We really need a more spectacular failure
when something like that happens.

The changes are:

 - chroma-agent.spec remove dependency on pcs (pacemaker/corosync)
   as the dependency is now a runtime not a install time one
 - manage_targets.py seperate the creation of the mount point into
   its own method and make the mount and unmount commands more
   resilient to the mount already being mounted or unmounted
 - chroma-bundles/Makefile add base_managed_no_ha_RH7.profile to
   live of profiles
 - base_managed_RH7.profile add pcs to the list of require packages
   as it is now a runtime dependency
 - base_managed_no_ha_RH7.profile copy of base_managed_RH7.profile
   with the pacemaker / corosync options set to false (not required)
 - filesystem.py bug fix and resilience of mount / unmount
 - filesystem_ldiskfs.py add resilience of mount / unmount
 - filesystem_zfs.p bug fix and resilience of mount / unmount
 - filesystem.py small refactor of UnmountStep usage
 - host.py small refactor of UnmountStep usage
 - target.py targets now have high_availibility property and
   through this property we can change the behavior ever so
   slightly at the point where targets are started/mounted,
   stopped/unmounted and created.
 - because tests now take advantage of the non-HA install as well
   as the HA install - test coverage is maintained.
 - Update fenche-agents from 4.0.11-27 to 4.0.11-47

Change-Id: I36f9124d8b557329fe5b18fea868bd91df182576
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: https://review.whamcloud.com/24546
Tested-by: Jenkins
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User