HYD-5210 Poll for target state rather than use a fixed delay
There is a test that has started failing intermittently
where an audit was expected to have come in carrying the
new state by 30 seconds later, but is not approx 30% of
the time. Specifically, it is checking that we detect that
a filesystem was unmounted. Since our scans are single
threaded, other activities in the scan taking a bit longer
can delay how long it takes to discover something like an
unmount.

Until we convert our scans to be multithreaded, we need to
loosen the restriction on how much time we wait for the
mount/unmounts to be detected, or we are slowed down by
intermittent test failures. Also, we should be using a
wait_until_true loop - no sense waiting around longer than
we have to if its detected correctly sooner!

  - replace time.sleep(30) in test_filesystem_detection.py with
    wait_until_true.
  - create a new function targets_in_state for waiting for all
    targets to achieve a specific state to reduce duplicate code.

Change-Id: I45b5e03878fcd5138d35b63f274f2acb2c67c2b7
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/23965
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Chris Gearing <chris.gearing@intel.com>