HYD-6887 Epel disable races with yum metadata expiry
The upgrade test tests the epel check in our installer by creating
an epel.repo, verifying the install fails, then re-enabling the check,
and verifying it passes. However, just removing the epel.repo file
doesn't remove epel from yums cached metadata, so there is a window
between when we remove the file and the yum metadata expires where
a yum command can actually still access the epel repo and cause the
verification it passes to fail. Normally our infrastructure has an
expiry of 1m so it is not an issue, but if it is increased, such as
to 15m, then the test can fail.

Our test script should not be vulnerable to variations to the metadata
expiry and should call yum-config-manager --disable epel to disable the
repo so the cache will also see it as disabled. Also, there is no reason
to be creating a new epel repo, we can just enable/disable the repo that
already should exist.

  - Add yum-config-manager --disable epel to
    tests/framework/integration/installation_and_upgrade/run_tests
  - Disable existing epel repo rather than creating a new one.

Change-Id: I3a0c9b5a324bae0d0309c5f9a0f8e4ae764bde15
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>