HYD-6885 Extend TEST_TIMEOUT
With the addition of managed ZFS, some activities on the manager are
expectedly taking a bit longer than they used to. This is causing some
intermittent test failures where commands don't complete by TEST_TIMEOUT,
but we can see in the test output complete from 3 to 120 seconds later.

There are some optimizations that could be made that can improve some of
these, such as the multithreaded agent change, but until these are
addressed in a future release, we are raising the value of TEST_TIMEOUT
so we don't lose the velocity to the intermittent test failures in the
meantime.

  - Change TEST_TIMEOUT from 5 * 60 to 10 * 60.

Change-Id: Ie0959a2b35046c7ec4b2384d7e49539f685d063c
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/24018
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>