HYD-6884 Refresh yum metadata after release --set
The yum clean all in upgrade_os is in the wrong place, it needs
to come after the subscription-manager release --set as the set
can change the repos available. Further, we need to do an explicit
makecache after the clean as the immediately following yum repolist
does not cause new metadata to then be fetched. We also need to
filter the exclamation point when searching for the epel repo name
in case the makecache takes longer than the yum metadata expiry.

Move 'yum clean all' to after 'subscription-manager release --set'
and add a 'yum makecache' after it, but before the yum repolist
used to disable the epel repo. Alter the regex to find the epel
repo name to exclude the potential exclamation point in the epel
repo name.

  - Make the above modifications to the commands in
    chroma-manager/tests/framework/integration/utils/upgrade_os.sh

Change-Id: I26fbe100302812e3eb2bd7392d7543dcc2594511
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/24019
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Chris Gearing <chris.gearing@intel.com>