HYD-6912 Agent log reports "no such pool" during zfs device full_scan()
Agent log shows stacktrace indicating failure when to trying to
export a zpool that is not currently imported. The failure occurs in
the agent linux plugin during
device_plugins/linux_components/zfs.py:_search_for_active().

The probable cause is that ExpirtedZfsDevices is not used when adding
the zpools and a lock is not acquired therefore enabling a race
condition to exist between different threads operating on the same
zpool.

This patch refactors zfs.py device plugin linux component to use
ExpiredZfsDevice when exporting or importing zpools. Therefore making
the code thread-safe.

  - Change _search_for_{,in}active() to return lists of zpool names
  - Change full_scan() to collect set of zpool names and add to
    block_devices within ExportedZfsDevice clause
  - Update test_zfs.py agent unit tests

Change-Id: I590afd38c63f603402006f45c744040313454e99
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: https://review.whamcloud.com/24221
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Chris Gearing <chris.gearing@intel.com>