HYD-6931 Test errors in FailoverTestCaseMixin cause issues with ZFS
create_filesystem_standard presumes that the volume ids passed to
create the file system will be the ones used by the files system. This
is not a safe assertion as the creation of a filesyste can create new
logical volumes that will then be used in the filesystem.

An example of this is the fact that Volume-Id's representing zpools are
passed for creation of the filesystem but the dataset created are the
VolumeIDs used in the filesystem.

Change create_filesystem_standard to read back the volume ids actually
used by the filesystem.

Another issue requiring fixing is that
_check_targets_for_volumes_started_on_expected_hosts has an if clause
in it meaning if unexpected data is found the test passes rather than
fails. This meant the when unknown id's were presented the test passed
causing the filesystem to failback before it had failedover.

TestFilesystemDNE was saving the volume ids for the FS as if they
might be used for future use. They are not used and so this saving was
a confusion and bug waiting to happen because they were not the id
actually used by the FS.

 - Cause create_filesystem_standard to read back the volumes used
 - Update FailoverTestCaseMixin to fail with a test error
 - Make _standard_filesystem_layout be created in the constructor as null
   and then be updated once the information is known. With an exception
   if it is used before it can be known
 - Remove storage of volume ids in TestFilesystemDNE so as not to mislead
   future developers

Change-Id: Ia1115ef2ab59ba6db4b0987e4d20e3bdc4737079
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: https://review.whamcloud.com/24305
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>