HYD-6938 don't use rsync in tests
Using rsync to copy files around the cluster in our tests doesn't really
add any value.  There was a comment saying that it was added in case the
network is slow but rsync doesn't add value in case the network is slow.
It only adds value if you need to run it multiple times to copy the same
content around.

What it does do is adds a requirement on nodes with no return value.

So remove it.

 - replace rsync with scp in chroma-manager/tests/framework/integration/shared_storage_configuration/full_cluster/cluster_setup

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Ieb5b0b995287223eb239be95ae54ba6dc8764ad0
Reviewed-on: https://review.whamcloud.com/24382
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>