HYD-6971 Chroma externals unncessarily cloned
The chroma-externals repository is being unncessarily cloned because the
update_chroma-externals.py always issues a 'git submodule init'. This is
not necessary and whilst generally harmless leads git to sometimes
reinitialize the whole repository.

This patch changes the behaviour to follow the git documentation and
only do a 'git submodule init' when no chroma-externals repository 
exists.

 - Add selective 'git submodule init'
 - Move the main code into a main function to remove warnings about
   duplicate variable names within nested scopes.

Change-Id: Ic1f2c1b990e05b5fcaf997b3a3d63eaae6def3c0
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: https://review.whamcloud.com/24389
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>