HYD-6945 TestWriteconf.testUpdateNids times out with single threaded agent
tWriteconf.testUpdateNids times when run with a single threaded agent.
4.0 has a multithreaded agent but when running with 3.1 the test will
fail and then complete ~60 seconds later.

This patch increases the timeout for the testUpdateNids to cure this
problem. This patch should not if possible make it to b4_0 but tbh.
managing that is probably more of a problem than the downside is worth.

 - Change timeout from TEST_TIMEOUT to LONG_TEST_TIMEOUT for wait_command in
   testUpdateNids

Change-Id: I8aec53acfb7a5a316b8575c1b0e5c561ec4fdd88
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: https://review.whamcloud.com/24430
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>