HYD-6972 Exit with error when fence-agents rpm for cluster-sim are not available
If the fence-agents for the cluster-sim build are not available the
build process does not fail, we just get an obscure error in

test_new_pdu_learns_outlet_states
test_powercycle_operation
test_off_on_operations
test_toggled_outlet_does_not_trigger_fencing_update

Tying the cause and effect together can be a long and vacation
destroying process!

This patch makes the build process exit if the rpms are not available. 
It may well be the case that others builds suffer the same fail quitely
behaviour but it is also probably the case that the knock on effect is
less subtle.

 - Add set -e to the build process for fence-agents
 - Update the version of fence-agents to 4.0.11-47

Change-Id: I4823bcc32e47453607de0f2f8dade07eb711419e
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: https://review.whamcloud.com/24740
Reviewed-by: Joe Grund <joe.grund@intel.com>
Tested-by: Joe Grund <joe.grund@intel.com>