HYD-6978 Update to Lustre build version 2.7.19.8
The Lustre team has provided us with a build for the most recent CVE, which
needs integrated into overall EE. They have also indicated that a new version
of dkms should be used with this build.

Update to the Lustre version 2.7.19.8 by changing the following build
parameters:

  JENKINS_BUILD_NUMBER = 165

Update to the dkms version dkms-2.3-1.20161202gitde1dca9.el7.noarch.rpm.

The following chroma-dependencies are effected:
  lustre[-client]
  zfs
  dkms

  - Update chroma-dependencies/include/Makefile.lustre_build_meta
  - Update chroma-dependencies/epel/Makefile

Change-Id: I9732de310b80cae54d2af9a6ddd751f1c6f1c08c
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/24695
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>