HYD-6977 Update to Lustre build version 2.5.42.28
The Lustre team has provided us with a build for the most recent CVE, which
needs integrated into overall EE. They have also indicated that a new version
of dkms should be used with this build.

Update to the Lustre version 2.5.42.28 by changing the following build
parameters:

  JENKINS_BUILD_NUMBER = 299

Update to the dkms version dkms-2.3-1.20161202gitde1dca9.el6.noarch.rpm.

The following chroma-dependencies are effected:
  lustre[-client]
  zfs
  dkms

  - Update chroma-dependencies/include/Makefile.lustre_build_meta
  - Update chroma-dependencies/epel/Makefile

Change-Id: I91aed56949939a481501da0c05cae39f33ff5733
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/24693
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>