HYD-7045 Incorrect kernel-devel version in managed profile
The base_managed_RH7.profile was not updated in the recent lustre
version to install the correct kernel-devel. Unexpectedly, this did
not cause test to fail, so it got by unnoticed.

This correctly bring this up to date, and now we know this can be
a silent failure. A separate patch will likely be needed to prevent
this from being silent as it is easy to miss.

  - Update kernel-devel version in base_managed_RH7.profile
  - Update simulator test reference to kernel-devel version

Change-Id: If1fc4d8238a788f7a510f8f5596c4f192bcb74ad
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/25078
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>