HYD-7071 Inconsistency with max_length param for Django CharFields
Max_length of django charfield at some point was changed to os.pathconf('.',
'PC_PATH_MAX'). This value seems to differ on OSX and Linux and causes problems
when creating db migration files. The lowest common denominator seen 
for this value is 1024, therefore this has been assigned as the default.

  - Create constant for charfield Max length in chroma_core/models/utils
  - Apply constant to max_field parameter of charfields instead of
    using os.pathconf()

Change-Id: I48110623f77618b2d2e230171a3ed33d4489dcb8
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: https://review.whamcloud.com/26808
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Peter Robinson <peter.timothyx.robinson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>