HYD-7096 Create a setting to disable power control device monitoring
IML has a hardware configuration requirement that the manager itself
be able to talk to the power control devices in order to monitor their
health. The manager itself directly monitors these.

AEON has an upcoming install where they've designed the solution so that
IPMI is done as a private network between pairs. This is for security -
if one node is compromised it cannot power down the entire cluster. This
prevents the manager from being able to monitor the BMCs, so IML is raising
alerts for being unable to reach them. Power control is still configured on
the lustre servers and HA should be operational, it's just the alerts that
are a problem. This really is a vali use case we need to support.

In the short term for this AEON install, we will implement a quick fix for a
setting to disable this monitoring so they don't have permanent alerts. The
tradeoff of changing this setting to True is that IML will then not be
monitoring the powe control devices for failure, and that if they fail
automatic failure will not occur, and manual intervention will be required
to restore the file system to service. The customer would be responsible for
setting up their own monitoring if they wanted to be informed of power control
device failure.

In the long term, we'd like to move the power control device monitoring to the
lustre servers that would actually be doing the stonithing. But that is out of
scope for this ticket in order to turn around on the schedule required for this
delivery.

  - Add new setting to disable power sontrol device monitoring.
  - If the setting is True, don't attempt to monitor the devices.

Change-Id: I25f8464935c6ccaf19ed4e7fee4fbd4793f67dec
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/25264
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>