HYD-7237 Reduce complexity of ExceptionSandbox
The ExceptionSandBox class will log errors when debug is off and suppress the
exception from being propagated. It is thought that this may be unnecessary and
that suppressing exceptions and printing summary information regarding sandboxed
exceptions to debug log (in both debug and normal modes) is a more appropriate
approach.

  - Remove ExceptionSandBox class, move logging and exception catching
    functionality in to exceptionSandBox function
  - Log a summarised traceback as debug when exception is thrown in sandbox
  - Remove references to ExceptionSandBox
  - Amend tests as necessary

Change-Id: Ibde4a698ac40561b33b67cfd6ac0a841d062378d