HYD-7145 Remove deprecated create_filesystem_simple test function
The test function create_filesystem_simple in the SSI tests has
been deprecated for some time in favor of the more robust
create_filesystem_standard. There are a few tests lingering that
are still using this test function. It would be good to clean up
the code and remove this deprecated function so we don't have to
continue to maintain it.

Remove the references to create_filesystem_simple and replace with
create_filesystem_standard, or a custome file system, as appropriate
for the individual test.

  - Remove all references to create_filesystem_simple
  - Remove create_filesystem_simple from chroma_integration_testcase
  - Remove, rather than update, test_autodetection, as it is
    redundant with our monitoring only (efs) tests.
  - Also fixes a little unrelated problem in remote_operations that was
    found while trying to debug this patch, where the test error message
    was not correctly rendering to a string.

Change-Id: I27226e80b583a9ae09a591d2ad11bc037445347a
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/25540
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>