HYD-7151 Device readonly check writes 0 length file in its place if device removed
Device scanning has a test to detect if a device is readonly or removed media that
does an open(device, 'w'). However, if there is a device where the path doesn't
exist at all anymore, such as a destroyed zpool, this will write a zero-length
file where the file is supposed to be. This causes intermittent failures in our
tests, such as missin 'part1' partitions for zpools, difficulty starting targets,
etc.

Instead, our code should use os.open(device, os.O_WRONLY), which will raise an OSError
if the path has disappeared, rather than creating a 0 length file. We then catch
the error and ignore it. Simply doing an os.path.exists prior is not sufficient, as
there would still be a small race where the device disappears after the check, but
prior to the open.

  - Change the open to use os.open with os.O_WRONLY
  - Add a comment warning against ever using 'w' with a regular open there
  - Close the file we're opening when we're done with it, which also appears to
    have been omitted.

Change-Id: I5bec7edc632ea6666ca456c7c598f0a9fd9829c6
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/25629
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>