HYD-7188 Gerrit search for a sha1 can return related changes also
When trying to find the change ref for updating chroma-externals, it seems
that gerrit will return changesets related by Change-Id: even if they don't
have the sha1 that is being queried.
This can lead to update_chroma-externals failing to find the right
changeref to update to.

Be more precise in finding the correct changeset

 - update update_chroma-externals.py to be more precise about which
   changeset it uses

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I9507259940e2d92180439561eea9aa03b2f7adf2
Reviewed-on: https://review.whamcloud.com/26299
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Peter Robinson <peter.timothyx.robinson@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>