HYD-7179 Failure in upgrade process does not end the upgrade test
It would seem that we are not propagating errors in upgrade_os() up to
it's caller and then not killing the test in the overall upgrade_os
test if the upgrade_os() function does have a failure in it.

This was noticed when upgrade_os() was trying to remove the O/S provided
yum repo files and they don't exist on RHEL yet this was not causing the
error to propagate up.  So might as well fix that also.

 - add error propagation to upgrade_os()
 - fail the upgrade_os test if upgrade_os() reports an error

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I539b3e27e42e667020a6561aedf252a8c803bd5e
Reviewed-on: https://review.whamcloud.com/25696
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
(cherry picked from commit 3760a269c0f8ec9ba7ec92168a6a7e589232a3cd)
Reviewed-on: https://review.whamcloud.com/26179
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>