HYD-7153 Ensure power control configured before file system creation in test
We seem to have skipped configuring power control in a number of tests over time.
This is not a supported configuration at this time, and we should ensure all
tests have power control configured as a user would.

Add a check to create_filesystem that power control has been configured for the
hosts the filesystem is on.

  - Add verify_power_control_configured to chroma_integration_testcase.py
  - Add call to verify_power_control_configured to create_filesystem
  - Update test cases that were failing to configure power control to do so
  - Cleaned up test_filesystem_same_name a bit now that create_filesystem doesn't
    make an assumption that all hosts added will have targets from the new
    filesystem.

Change-Id: I0e36eafc8ba538e8da7c030ad36771980b4d8977
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/25902
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>