HYD-7230 Work around RH Bug 1430112
There is a race when Pacemaker starts up that it may decide it
needs to fence the other node. However, if this is the first
time the cluster is coming up, there will be no fencing device
yet, so then this fencing attempt fails. Pacemaker then
"remembers" this failure. This has a knock on effect that
subsequent crm_resource commands might take several minutes.
When this happens, the step initiating the crm_resource command
can timeout and fail. The current mechanism is that if no change
is detected in 2 minutes, we timeout, and we've observed it take
more than 15 minutes in this scenario.

The folks at RH are working on fixing it proper in RHEL 7.4, but in the meantime, in the event of a timeout, retry as a temporary work around to this pacemaker bug.

  - Add retry to stop actions in
    chroma-agent/chroma_agent/action_plugins/manage_targets.py

Change-Id: I835c48c16d96dd7085d803c996d748f455d08fa9
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/26407
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>