HYD-7190 O/S upgrade failing on RHEL 7.2
On RHEL 7.2 the subscription-manager release --set can fail to
happen silently.  This is documented upstream in
https://bugzilla.redhat.com/show_bug.cgi?id=1428435.

That ticket describes a workaround to stop rhsmcertd if it is
hung and then update the repos.

But the reason that rhsmcertd-worker was hanging was because
rhsmd was not being started with proxy configuration.  That
wasfixed in the provisioner.

What's left here is to ensure that the upgrade happens.

 - update upgrade_os.sh to harden the upgrade to ensure
   that it was done

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I95ad2dd896d28c7819672c6d4dba14108e661dc8
Reviewed-on: https://review.whamcloud.com/26080
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User