HYD-6200 Reduce copious stdout
To eliminate the "error: Resource temporarily unavailable" errors
we get.
This is actually being caused by various tools not handling EAGAIN
on their stdout that is set to non-blocking mode.
This is not normally a problem when writing to stdout as it's usually
a file or tty, but when it's a network socket that can block, like a
Jenkins' slave's stdout would be, the writer could get an EAGAIN while
trying to write it's stdout to the Jenkins log over the network.

The tools should be fixed.

But in the meanwhile let's just reduce the copiousness of some output
(which is probably a good thing anyway) and see if we an mitigate that
way.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I81f19b281b304237f499a7ed5e4527f0701a7f8a
Reviewed-on: https://review.whamcloud.com/26325
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>