HYD-7298 Use EPEL as defacto dependency source
We don't want the overhead of maintaining our own SRPMs, and building
IML from scratch.

Pull as many dependencies as possible from EPEL.

This involves converting over all necessary dependencies and then
auditing what's left and trying to convert any code needed so it is
compatible with EPEL versions.

The end goal is to have as little custom packaged source as possible.

Exceptions that need to be worked in separate tickets:
 - python-django
 - python-jsonschema

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I9a9512d1203d548333d73ee9399a7d7756ae6e30