HYD-6900 Retry on failure to get fqdn
It seems that occasionally, python's socket.getfqdn() can fail to
return a value when simply calling it again, it succeeds.  Maybe
this is a Python/glibc bug, maybe this is an infrastructure problem.

Either way, be more resilient to this failure and try again.

Various other fixes in this patch to comply with PEP-8.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Ie9a4fe2eaef9fbaf63b6a6db7d6ab54e36de58b8
Reviewed-on: https://review.whamcloud.com/26720
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>