HYD-7314 Display results of fencing configuration in test
In our testing, we seem to find instances where a fencing device
cannot be found such as:

notice: Couldn't find anyone to fence (reboot) <host> with any device

It is strange that Pacemaker would not be able to find a fencing device
if it was configured properly.

Let's print out some results of the fencing configuration to show that
all nodes are fencable after we configure fencing in our testing frame-
work.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I63ac349ad9e0a567408a935b6de7747e826379d7
Reviewed-on: https://review.whamcloud.com/26747
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>