HYD-7322 Fix races between test_filesystem_same_name and device discovery
test_filesystem_same_name has some races where it waits for changes to
appear on 2 servers as it creates a filesystem on 2 servers. But the
test functions don't allow you to specify which servers, so there is a
race here where it might be discovered on the ones we want, or it might
be a different subset of the total 4 servers.

Ultimately, it would be good for the shared test functions to allow you
to specify which of the servers the devices need to be discovered on,
but for now we can address an intermittent test failure quickly by just
waiting for the volumes to be discovered on all 4 servers instead of just
the 2.

  - Wait for devices to be detected on all 4 servers instead of just 2.

Change-Id: Ifb9cff514f0c4b0ffd5f7e49c9d13bfb0d5b9b6f
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/26822
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>