HYD-7323 Don't match a volume_node to a target if host is None
Prior to MakeTargetActive step, we have some logic to select the
active_volume_node for a target. However, it has a logic error where
if host is None and it encounters a VolumeNode where volume_node.host
is also None, it could end up assigning an active_volume_node. This
then fails the assert a few lines later that makes sure
active_volume_node is set if and only if host is not None. This causes
an occasional intermittent test failure.

Add a check to the logic that looks for a matching node to only set
and active_volume_node if host is not None.

  - Add host is not None check to create_parameters.

Change-Id: I51cc0162333950acb41bb21c8898011c07346c45
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/26824
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Peter Robinson <peter.timothyx.robinson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>