HYD-7336 Make zfs import and export idempotent
The way zpool import/export is implemented currently, if any operation
occurs on a zpool outside of the zpool locking mechanism (including a
user action), there can be a little race between when we check if the
pool is imported before deciding to import/export appropriately.

If we use a try/except model, rather than a if/then model, we can
eliminate that potential race. This also actually really simplifies the
set of possible code paths, as you can see by the resulting
modifications to the unit tests.

  - update BlockDeviceZfs import_ and export to go ahead and just try
    to import/export, and pass on if it finds it was already in the
    correct state.

Change-Id: I62c4bf80903ae3ac6ac6ffddd370525c26849318
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: https://review.whamcloud.com/26826
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>