HYD-7326 Don't assume the dirname of the code checkout
The directory name a git repo is cloned into by default is named
the same as the project.  This makes any code that references
such a dir dependent on the project name.

Have the caller tell the test code where the code checkout is.

Also update the default to be coherent with the current project
name