HYD-7338 Fix exceptions due to missing config store directories
During agent daemon start-up, the settings directory of the config
store sometimes cannot be found and results in an exception in the
chroma-agent damon log.

The explanation for this is that the chroma-agent service has been
started before agent registration (in agent_setup.py) has completed
(during which the config store directories are created).

This situation can occur due to a bug when checking if the service
is already running during chroma-agent registration. After the
fix is applied, if this situation occurs, a message should appear
in the agent daemon log indicating "service was running before
registration, stopping.".

  - stop agent service if it is already running at the beginning
    of the agent_setup.py 'register' function
  - verfiy both section and key exist in config_store before
    continuing with agent main loop

Change-Id: I8df2e20cbdd25c0e517150546283d369e028082b